app/src/main/java/com/example/dagger2/login/LoginViewModel.kt
package com.example.dagger2.login

import android.os.UserManager
import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import javax.inject.Inject

/**
 * LoginViewModel is the ViewModel that [LoginActivity] uses to
 * obtain information of what to show on the screen and handle complex logic.
 */

//В этом случае LoginViewModel не нужно повторно использовать другими классами,
// поэтому мы не должны аннотировать его с помощью @ActivityScope.
class LoginViewModel @Inject constructor(private val userManager: com.example.usage.UserManager) {

    private val _loginState = MutableLiveData<LoginViewState>()
    val loginState: LiveData<LoginViewState>
        get() = _loginState

    fun login(username: String, password: String) {
        if (userManager.loginUser(username, password)) {
            _loginState.value = LoginSuccess
        } else {
            _loginState.value = LoginError
        }
    }

    fun unregister() {
        userManager.unregister()
    }

    fun getUsername(): String = userManager.username
}
